moved optimize_scene.py removed f_strings
